--- src/main/java/com/coding_bottle/Calendar/CalendarApplication.java
@@ -1,13 +1,0 @@
-package com.coding_bottle.Calendar;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CalendarApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(CalendarApplication.class, args);
-	}
-
-}
