--- src/main/java/com/coding_bottle/Calendar/domain/auth/dto/LoginDto.java
@@ -1,7 +1,0 @@
-package com.coding_bottle.Calendar.domain.auth.dto;
-
-public record LoginDto(
-        String username,
-        String password
-) {
-}
