--- src/main/java/com/coding_bottle/Calendar/domain/member/dto/MemberResponseDto.java
@@ -1,8 +1,0 @@
-package com.coding_bottle.Calendar.domain.member.dto;
-
-public record MemberResponseDto(
-        Long memberId,
-        String email,
-        String nickname
-) {
-}
