--- src/main/java/com/coding_bottle/Calendar/domain/member/repository/MemberRepository.java
@@ -1,10 +1,0 @@
-package com.coding_bottle.Calendar.domain.member.repository;
-
-import com.coding_bottle.Calendar.domain.member.entity.Member;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    Optional<Member> findByEmail(String username);
-}
