+++ src/main/java/com/codingbottle/calendar/CalendarApplication.java
@@ -1,0 +1,13 @@
+package com.codingbottle.calendar;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CalendarApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(CalendarApplication.class, args);
+	}
+
+}
