+++ src/main/java/com/codingbottle/calendar/domain/auth/dto/LoginDto.java
@@ -1,0 +1,7 @@
+package com.codingbottle.calendar.domain.auth.dto;
+
+public record LoginDto(
+        String username,
+        String password
+) {
+}
