+++ src/main/java/com/codingbottle/calendar/domain/member/dto/MemberResponseDto.java
@@ -1,0 +1,8 @@
+package com.codingbottle.calendar.domain.member.dto;
+
+public record MemberResponseDto(
+        Long memberId,
+        String email,
+        String nickname
+) {
+}
