+++ src/main/java/com/codingbottle/calendar/domain/member/repository/MemberRepository.java
@@ -1,0 +1,10 @@
+package com.codingbottle.calendar.domain.member.repository;
+
+import com.codingbottle.calendar.domain.member.entity.Member;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    Optional<Member> findByEmail(String username);
+}
