--- src/test/java/com/coding_bottle/Calendar/CalendarApplicationTests.java
@@ -1,13 +1,0 @@
-package com.coding_bottle.Calendar;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-// @SpringBootTest
-class CalendarApplicationTests {
-
-	// @Test
-	void contextLoads() {
-	}
-
-}
