+++ src/test/java/com/codingbottle/calendar/CalendarApplicationTests.java
@@ -1,0 +1,10 @@
+package com.codingbottle.calendar;
+
+// @SpringBootTest
+class CalendarApplicationTests {
+
+	// @Test
+	void contextLoads() {
+	}
+
+}
